AI-Powered-Trading-Bot-main/requirements.txt
pandas
matplotlib
python-binance
pandas_ta
scikit-learn
python-dotenv
uvicorn
fastapi
pydantic
websockets